Move read_station_metadata() to its own file
